tests/wpt/tests/css/css-logical/parsing/border-inline-style-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: parsing border-inline-style with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-border-block">
<meta name="assert" content="border-inline-style supports the full grammar '<line-style>{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// none | hidden | dotted | dashed | solid | double | groove | ridge | inset | outset
test_valid_value("border-inline-start-style", "dotted");
test_valid_value("border-inline-start-style", "groove");
test_valid_value("border-inline-start-style", "inset");
test_valid_value("border-inline-start-style", "none");
test_valid_value("border-inline-start-style", "solid");
test_valid_value("border-inline-end-style", "dashed");
test_valid_value("border-inline-end-style", "double");
test_valid_value("border-inline-end-style", "hidden");
test_valid_value("border-inline-end-style", "outset");
test_valid_value("border-inline-end-style", "ridge");
test_valid_value("border-inline-style", "dotted");
test_valid_value("border-inline-style", "double groove");
test_valid_value("border-inline-style", "hidden hidden", "hidden");
</script>
</body>
</html>
